Code refactor and bug fixes.
