+++ src/main/java/com/softserve/skillscope/talent/model/request/TalentEditRequest.java
@@ -1,7 +1,6 @@
 package com.softserve.skillscope.talent.model.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.Builder;
 import org.hibernate.validator.constraints.URL;
@@ -10,13 +9,10 @@ import java.time.LocalDate;
 
 @Builder
 public record TalentEditRequest(
-        @NotBlank(message = "Password mustn't be blank ")
         @Size(min = 5, max = 64, message = "Password must be less than 64 characters and more than 5")
         String password,
-        @NotBlank(message = "Name mustn't be blank ")
         @Size(min = 1, max = 64, message = "Name must be less than 64 characters and more than 1")
         String name,
-        @NotBlank(message = "Surname mustn't be blank ")
         @Size(min = 1, max = 64, message = "Surname must be less than 64 characters and more than 1")
         String surname,
         @URL
